requirements.txt
beautifulsoup4==4.5.1
dj-database-url==0.4.1
Django==1.10.1
django-admin-bootstrapped==2.5.7
django-authtools==1.5.0
django-blog-zinnia==0.18.1
django-braces==1.9.0
django-contrib-comments==1.7.3
django-crispy-forms==1.6.0
django-environ==0.4.0
django-mptt==0.8.6
django-tagging==0.4.5
django-xmlrpc==0.1.5
easy-thumbnails==2.3
gunicorn==19.6.0
mots-vides==2015.5.11
Pillow==3.3.1
pyparsing==2.1.9
pytz==2016.6.1
regex==2016.9.22
requests==2.9.1
six==1.10.0
whitenoise==3.2.2
psycopg2==2.5.4